adapt tests for the streaming branch

+++ NewTests/InterruptedDownloads/test_interrupted_downloads.py
@@ -81,7 +81,7 @@ class TestInterruptedDownloads(OrthancTestCase):
     def test_interrupting_a_download_should_cancel_the_archive_job(self):
         
         # note: there was a regression between 1.12.10 and 1.12.11
-        if self.o.is_orthanc_version_at_least(1, 12, 12):
+        if self.o.is_orthanc_version_at_least(1, 12, 12) and self.o.get_system()["ApiVersion"] >= 32:  # 1.12.11+ and 32 = streaming branch
             self.o.delete_all_content()
             populator = OrthancTestDbPopulator(self.o, studies_count=1, series_count=2, instances_count=200, random_seed=65)
             populator.execute()
